move colorscale from js_plotting_utils to _engine_utils
