+++ nilearn/plotting/tests/test_html_stat_map.py
@@ -9,6 +9,7 @@ from nibabel import Nifti1Image
 
 from nilearn import datasets, image
 from nilearn.image import get_data, new_img_like
+from nilearn.plotting._engine_utils import colorscale
 from nilearn.plotting.html_stat_map import (
     StatMapView,
     _bytes_io_to_base64,
@@ -27,7 +28,6 @@ from nilearn.plotting.html_stat_map import (
     _threshold_data,
     view_img,
 )
-from nilearn.plotting.js_plotting_utils import colorscale
 
 
 def _check_html(html_view, title=None):
